pyproject.toml
[tool.poetry]
name = "gen-tax-report"
version = "0.1.0"
description = ""
authors = ["Chaipat Srikajohnlap <gorufu666@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
aiohttp = "^3.8.4"
requests = "^2.31.0"
pandas = "^2.0.3"
python-dotenv = "^1.0.0"
apscheduler = "^3.10.1"
orjson = "^3.9.2"
tortoise-orm = "^0.19.3"
arrow = "^1.2.3"
oauth2client = "^4.1.3"
gspread = "^5.10.0"
cryptography = "^41.0.3"
aiomysql = "^0.2.0"
tqdm = "^4.66.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
